+++ web/app/themes/brocooly/bootstrap/config.php
@@ -20,3 +20,12 @@ if ( ! defined( 'BROCOOLY_THEME_PATH' ) ) {
 if ( ! defined( 'BROCOOLY_THEME_URI' ) ) {
 	define( 'BROCOOLY_THEME_URI', trailingslashit( get_template_directory_uri() ) );
 }
+
+/**
+ * Path where all i18n files are
+ *
+ * @since 1.5.0
+ */
+if ( ! defined( 'BROCOOLY_THEME_LANG_PATH' ) ) {
+	define( 'BROCOOLY_THEME_LANG_PATH', BROCOOLY_THEME_PATH . 'languages' );
+}